[#2246] Split User from Users file
